% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateBulkRegion.R
\name{annotateBulkRegion}
\alias{annotateBulkRegion}
\title{Assign regions defined by bulk sequencing with cell types}
\usage{
annotateBulkRegion(
  cellGraph,
  labelEdges,
  labelEdges2,
  groups1 = 1,
  groups2 = 0,
  regionMat = NULL,
  ATACMatList = NULL,
  peaksList = NULL,
  labelEdgeWeights = c(1, 0.01),
  tr1 = NULL,
  wtree = c(1, 1),
  sampleDepth = 5000,
  compute.Zscore = T,
  nround = 50,
  ...
)
}
\arguments{
\item{cellGraph}{cell-cell graph, a cell-to-cell similarity matrix, row and colnames are cell barcodes}

\item{labelEdges}{a matrix or dataframe of edges from each cell type label to each cell. Each row is a cell,
each column is a cell type label
#' @param labelEdges2  a matrix or dataframe of edges from each bulk label to each cell. Each row is a cell,
each column is a bulk label}

\item{groups1}{a numeric vector showing groups of cells for permutation between cells and cell type labels. If a single number, 0 means no permutation, non-zero means permutation for all cells.
If a vector, must have the same length as the rows of \code{labelEdge}. Only permute edges within the same group of the cells. For cells with entry 0, no permutation.}

\item{groups2}{a numeric vector showing groups of regions for permutation between regions and bulk annotations .If a single number, 0 means no permutation, non-zero means permutation for all regions.
If a vector, must have the same length as the rows of \code{regionMat}. Only permute edges within the same group of the regions. For regions with entry 0, no permutation.
if both \code{groups1} and \code{groups2} are NULL, permute among all cells. If either of them is NULL, assume no permutation for the null object.}

\item{regionMat}{a data.table or dataframe with sequence name (chrX:XXX-XXX) as the first column and clusters as the following columns. The second to the last columns
are binary indicating whether a region belongs to a cluster.}

\item{ATACMatList}{a list of a peak by cell data.frame or matrix, colnames are cell barcodes. Must provided if permute bulk annotation to regions.}

\item{peaksList}{a list of a dataframe of genomic coordinates of ATACSeq data. Must contain at least three columns
with names 'seqnames', 'start' and 'end'. Must provided if \code{ATACMatList} is not NULL.}

\item{labelEdgeWeights}{the edge weight ratio between cell-cell type edges and cell-cell edges and between cell-bulk label edges and cell-cell edges.
If it's NULL, will tune edge weights and get the optimal value of it; otherwise will use the input weight}

\item{tr1}{if not null, incorporating the cell type hierarchy into the graph. input the cell type tree must be a phylo object}

\item{wtree}{the edge weights traversing up/down the cell type tree. default: 1}

\item{sampleDepth}{subsample cells to for faster calculation to tune edge weight ratio, default: 5000}

\item{compute.Zscore}{whether to do permutation and compute Z-score, If not, will return influence score between labels.}

\item{nround}{rounds of permutations to compute null distribution. default: 50}

\item{...}{other arguments pass to \code{tuneEdgeWeights} and \code{walkCells}}
}
\value{
A cellWalk2 object. if \code{compute.Zscore} is False a list including influence score from bulk labels to cell types and labelEdgeWeights.
otherwise a list including influence and Zscore from bulk labels to cell types, as well as labelEdgeWeights.
}
\description{
\code{annotateBulkRegion} Random walk on the cells, cell type labels and bulk labels graph and compute Z-scores from bulk labels to each cell type label.
Cell type labels can be organized in a hierarchical structure.
}
