% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCellTypesFunctions.R
\name{annotateCells}
\alias{annotateCells}
\title{Annotate cells by cell types
\code{annotateCells} Random walk on the cells and cell type labels graph and compute influence scores from each cell to each cell type labels.
Cell type labels can be organized in a hierarchical structure.}
\usage{
annotateCells(
  cellGraph,
  labelEdges,
  weight1 = NULL,
  sampleDepth = 3000,
  ...,
  tr1 = NULL,
  wtree = 1
)
}
\arguments{
\item{cellGraph}{cell-cell graph, a cell-to-cell similarity matrix, row and colnames are cell barcodes}

\item{labelEdges}{a matrix or dataframe of edges from each cell type label to each cell. Each row is a cell,
each column is a cell type label}

\item{weight1}{the edge weight ratio between cell-label edges and cell-cell edges.
If it's NULL, will tune edge weights and get the optimal value of it; otherwise will use the input weight}

\item{sampleDepth}{subsample cells to for faster calculation to tune edge weight ratio, default: 3000}

\item{...}{other arguments pass to \code{tuneEdgeWeights} and \code{walkCells}}

\item{tr1}{if not null, incorporating the cell type hierarchy into the graph. input the cell type tree as a phylo object}

\item{wtree}{the edge weight between cell type labels, default: 1}
}
\value{
a list including a CellWalker object and weight1
}
\description{
Annotate cells by cell types
\code{annotateCells} Random walk on the cells and cell type labels graph and compute influence scores from each cell to each cell type labels.
Cell type labels can be organized in a hierarchical structure.
}
