% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ATACMat}
\alias{ATACMat}
\title{Cell-by-peak matrix}
\format{
A matrix with 1723 rows and 100000 columns
}
\usage{
ATACMat
}
\description{
A cell-by-peak matrix generated from a scATAC-seq experiment
}
\keyword{datasets}
