% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellWalkerFunctions.R
\name{combineGraph}
\alias{combineGraph}
\title{Generate combined graph}
\usage{
combineGraph(labelEdges, cellEdges, weight = 1)
}
\arguments{
\item{labelEdges}{matrix with dimensions c-by-l}

\item{cellEdges}{matrix with dimensions c-by-c}

\item{weight}{numeric indicating label edge weight}
}
\value{
combined matrix with dimensions (c+l)-by-(c+l)
}
\description{
\code{combineGraph()} combines weighted cell-to-label edges (dimensions
c-by-l) with cell-to-cell edges (dimensions c-by-c) into a single large graph
}
\examples{
data("SampleCellWalkRData")
combineGraph(SampleCellWalkRData$labelEdges, SampleCellWalkRData$cellEdges)

}
