% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellWalkerFunctions.R
\name{combineMultiLabelGraph}
\alias{combineMultiLabelGraph}
\title{Generate combined graph with list of label edges}
\usage{
combineMultiLabelGraph(labelEdgesList, cellEdges, weights)
}
\arguments{
\item{labelEdgesList}{list of matrices each with dimensions c-by-l}

\item{cellEdges}{matrix with dimensions c-by-c}

\item{weights}{numeric vector indicating label edge weight for each label edge in list}
}
\value{
combined matrix with dimensions (c+L)-by-(c+L) where L is the sum of l's
}
\description{
\code{combineMultiLabelGraph()} combines list of weighted label edges with
cell edges into a single graph
}
