% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateBulkRegion.R
\name{computeBulkEdges}
\alias{computeBulkEdges}
\title{Compute Label Edges}
\usage{
computeBulkEdges(regions, peaks, ATAC_Mat, drop = T)
}
\arguments{
\item{regions}{a dataframe of genomic coordinates of regions. Must contain either four columns
with names 'seqnames', 'start', 'end' and 'cluster' or two columns: 'sequence_name' and 'cluster'. Can have an additional column named 'weight' to weight each region.}

\item{peaks}{a dataframe of genomic coordinates of ATACSeq data. Must contain at least three columns
with names 'seqnames', 'start' and 'end'}

\item{ATAC_Mat}{a peak by cell data.frame or matrix, colnames are cell barcodes}

\item{drop}{whether to drop the labels with no overlapping with regions}
}
\value{
a matrix of edges from each bulk label to each cell
}
\description{
\code{computeBulkEdges} generates a matrix of edges from each bulk label (e.g. a group of genomic regions) to each cell by ATACSeq data.
The edge weight is proportion of read counts in each set of regions.
}
