% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellWalkerFunctions.R
\name{computeCellHomogeneity}
\alias{computeCellHomogeneity}
\title{Compute cell homogeneity}
\usage{
computeCellHomogeneity(cellWalk, cellTypes)
}
\arguments{
\item{cellWalk}{a cellWalk object}

\item{cellTypes}{list of names of cell types, if not provided unique labels are used}
}
\value{
cell homogeneity
}
\description{
\code{computeCellHomogeneity()} computes the median influence between cells of the same
type compared to cells of other types. This serves as a proxy for for how well cells
are labeled and can be used to tune parameters such as label edge weight
}
