% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{computeLabelEdges}
\alias{computeLabelEdges}
\title{Compute Label Edges}
\usage{
computeLabelEdges(
  labelGenes,
  ATACMat,
  ATACGenePeak,
  method = "Expression",
  filters,
  filterWeights,
  filterOut,
  filterGene,
  regions,
  peaks,
  whichMat = "bmat"
)
}
\arguments{
\item{labelGenes}{data.frame with genes of interest in first column and corresponding
labels in second column, optionally log-fold change in expression values in the third}

\item{ATACMat}{either a cell-by-peak matrix, a SnapATAC object, or an ArchR project}

\item{ATACGenePeak}{per label mapping of peaks to genes returned by mapPeaksToGenes()}

\item{method}{scaling method, either "Expression","Correlation", or "None"}

\item{filters}{list of GRanges lists for locations that pass filters}

\item{filterWeights}{numeric vector of weights assigned to each filter}

\item{filterOut}{boolean for each filter of whether those regions are permitted or filtered out}

\item{filterGene}{boolean for each filter of whether it applies to genes as a whole or just overlapping peaks (note that weights do not apply to peaks)}

\item{regions}{regions map peaks to genes}

\item{peaks}{GRanges of peaks in ATACMat}

\item{whichMat}{string determing whether to use "bmat" or "gmat" from snap object or "TileMatrix" or "GeneScoreMatrix" from ArchR project}
}
\value{
matrix of weights from each label to each cell
}
\description{
\code{computeLabelEdges} generates a matrix of edges from each label to each cell.
Edges are scaled by the passed weights and filters. Each filter can either apply
just to specific peaks or to whole genes (filterGene, TRUE by default) and can be permissive or
filtering out (filterOut, FALSE by default). Regions mapping peaks to genes must be
provided to use filters. If filters are applied to specific peaks, peaks for the ATAC
data must also be provided.
}
\examples{
data("SampleCellWalkRData")
computeLabelEdges(SampleCellWalkRData$labelGenes,
                  SampleCellWalkRData$ATACMat,
                  SampleCellWalkRData$ATACGenePeak)

}
