% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{computeMST}
\alias{computeMST}
\title{Comptute Minimum Spanning Tree}
\usage{
computeMST(
  cellWalk,
  cellTypes,
  labelThreshold,
  recompute = FALSE,
  plot = TRUE,
  seed
)
}
\arguments{
\item{cellWalk}{a cellWalk object}

\item{cellTypes}{character, optional vector of labels to use, all labels used by default. Most likely label among those listed
will be used. If only a single label is provided, all cells will be colored by their score for that label. If two labels are
given, the difference in score for each cell is shown.}

\item{labelThreshold}{numeric, set a threshold below which cells aren't labeled (e.g. 0)}

\item{recompute}{boolean, recompute layout for MST plot}

\item{plot}{boolean, optionally don't plot output and only compute the embedding}

\item{seed}{numeric, random seed}
}
\value{
cellWalk object with MST stored in "MST" and layout stored in "MST_layout"
}
\description{
\code{computeMST()} computes a minimum spanning tree based on cell-to-cell influence
}
\examples{
data("SampleCellWalkRData")
cellWalk <- computeMST(SampleCellWalkRData$cellWalk, plot=FALSE)

}
