% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{computeTFexp}
\alias{computeTFexp}
\title{Compute mean gene expression of TF in each cell type}
\usage{
computeTFexp(
  tf_exp,
  Zscore,
  cellLabels,
  tr = NULL,
  levels = c(1:8, 12),
  scale = T
)
}
\arguments{
\item{tf_exp}{single cell level gene expression of TFs}

\item{Zscore}{a matrix or dataframe, rows are TFs to be mapped to, and columns are cell types}

\item{cellLabels}{cell labels}

\item{tr}{If tr is provided, also compute selected internal nodes of the cell type tree.}

\item{levels}{select levels of internal nodes to compute expression}

\item{scale}{If scale = T, get standardized expression for each gene}
}
\value{
a list with Z-score and mean expression
}
\description{
\code{computeTFexp} Compute mean gene expression of TF in each cell type
}
