% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCellTypesFunctions.R
\name{computeTypeEdges}
\alias{computeTypeEdges}
\title{Compute Label Edges}
\usage{
computeTypeEdges(
  exprMat_norm,
  markers,
  pval.cutoff = 0.05,
  log2FC.cutoff = 0.5,
  only.pos = F,
  force = T
)
}
\arguments{
\item{exprMat_norm}{a gene by cell data.frame or matrix, rownames are genes, colnames are cell barcodes}

\item{markers}{marker genes for each cell type, columns are gene, cluster, p_val_adj (optional) and avg_log2FC (optional).}

\item{pval.cutoff}{select markers with adjust pvalue (\code{p_val_adj}) < pval.cutoff. Default: 0.05}

\item{log2FC.cutoff}{select markers with abs(log2FC.cutoff)> 0.5 if only.pos = F or log2FC.cutoff>0.5 if only.pos = T. Default: 0.5}

\item{only.pos}{only include positive markers}

\item{force}{if true (default), this function will use marker genes that matched the name in exprMat_norm to compute edge weight;
if false, this function will raise an error if some marker genes are not in exprMat_norm.}
}
\value{
a matrix of edges from each cell type label to each cell
}
\description{
\code{computeTypeEdges} generates a matrix of edges from each cell type label to each cell from gene expression.
The edge weight is normalized gene expression of markers weighted by the log2FC.
}
