% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctionsExtra.R
\name{constructCellGraph}
\alias{constructCellGraph}
\title{Construct cell-to-cell graph from multiple sources}
\usage{
constructCellGraph(
  RNA_Mat,
  ATAC_Mat,
  peaks,
  RNA_Mat2 = NULL,
  ATAC_Mat2 = NULL,
  peaks2 = NULL,
  distan = "sparseCosine",
  filterGene = T,
  ext_region = 10000,
  ATAC_weight = 0.3,
  ndim = 1:30,
  filterFreq = c(0.002, 0.2),
  logarithm = T,
  normalized = T,
  knn = 30
)
}
\arguments{
\item{RNA_Mat}{a gene-by-cell count matrix, RNASeq part of the multiomic data.Rownames are genes and colnames are cells}

\item{ATAC_Mat}{a peak-by-cell count matrix, ATACSeq part of the multiomic data. Must have barcodes as colnames}

\item{peaks}{a dataframe of genomic coordinates of peaks of ATACSeq part of the multiomic data. Must contain at least three columns
with names 'seqnames', 'start' and 'end'.}

\item{RNA_Mat2}{a gene-by-cell count matrix, unpaired RNASeq data.Rownames are genes and colnames are cells}

\item{ATAC_Mat2}{a peak-by-cell count matrix,unpaired ATACSeq data. Must have cell barcodes as colnames}

\item{peaks2}{a dataframe of genomic coordinates of peaks of unpaired ATACSeq data. Must contain at least three columns
with names 'seqnames', 'start' and 'end'. Must input peaks2 if \code{ATAC_mat2} is not NULL}

\item{distan}{distance metrics for compute similarity of ATACSeq profile. Either sparseJaccard, sparseCosine or LSI.}

\item{filterGene}{filter peaks close to gene body and promoters to compute cell-cell similarity}

\item{ext_region}{expand gene region by \code{ext_region} bp for filtering peaks. Default: 1e4}

\item{ATAC_weight}{the weight of ATACSeq similarity for combining ATACSeq and RNASeq profile of multiomic data. \code{ATAC_weight} ATAC similarity +
(1-\code{ATAC_weight}) RNA similarity. Must to between 0 and 1.}

\item{ndim}{the PC dimensions to compute cell-cell distance using RNASeq data}

\item{filterFreq}{filter peaks by proportion of cells having the peaks. a vector of low and high proportion cutoffs}

\item{logarithm}{Whether to take logarithm of the cell-cell distance using ATACSeq data}

\item{normalized}{Whether to normalize the cell-cell distances between 0 and 1}

\item{knn}{Defines k for the k-nearest neighbor algorithm}

\item{...}{additional parameters used by method}
}
\value{
matrix of cell-to-cell similarity
}
\description{
\code{constructCellGraph} construct cell-to-cell graph integrating RNASeq, ATACSeq and multiomic data.
}
