% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctionsExtra.R
\name{constructCellGraphFromATAC}
\alias{constructCellGraphFromATAC}
\title{Construct cell similarity graph based on ATACSeq data}
\usage{
constructCellGraphFromATAC(
  ATAC_Mat,
  peaks,
  distan = "sparseCosine",
  filterGene = T,
  ext_region = 10000,
  filterFreq = c(0.002, 0.2),
  logarithm = T,
  normalized = T
)
}
\arguments{
\item{ATAC_Mat}{a peak-by-cell count matrix. Must have barcodes as colnames}

\item{peaks}{a dataframe or GRanges object of genomic coordinates of peaks. If it's a dataframe, uust contain at least three columns
with names 'seqnames', 'start' and 'end'}

\item{distan}{distance metrics for compute similarity of ATACSeq profile. Either sparseJaccard, sparseCosine or LSI.}

\item{filterGene}{filter peaks close to gene body and promoters to compute cell-cell similarity}

\item{ext_region}{expand gene region by \code{ext_region} bp for filtering peaks. Default: 1e4}

\item{filterFreq}{filter peaks by proportion of cells having the peaks. a vector of low and high proportion cutoffs}

\item{logarithm}{Whether to take logarithm of the cell-cell distance using ATACSeq data}

\item{normalized}{Whether to normalize the cell-cell distances between 0 and 1}
}
\value{
matrix of cell-to-cell similarity
}
\description{
\code{constructCellGraphFromATAC} construct cell-to-cell graph from ATACSeq count data.
}
