% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{convert2plot}
\alias{convert2plot}
\title{Plot Z-score heatmap with gene expression}
\usage{
convert2plot(
  Zscore,
  tf_exp_mean,
  label = "zscore",
  th = 3,
  ord2 = NULL,
  ord = NULL
)
}
\arguments{
\item{Zscore}{a matrix or dataframe, rows are TFs to be mapped to, and columns are cell types}

\item{tf_exp_mean}{gene expression of TFs}

\item{label}{figure legend, default: zscore}

\item{th}{threshold of Zscore to plot}

\item{ord2}{order of cell types, if NULL, clustering cell types by Z-score}

\item{ord}{order of TFs, if NULL, clustering TFs by Z-score}
}
\value{
a list with reordered Z-score and expression to plot
}
\description{
\code{convert2plot} Plot Z-score heatmap with gene expression
}
