% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateBulkRegion.R
\name{convertToMatrix}
\alias{convertToMatrix}
\title{Convert bulk annotation dataframe to a binary matrix}
\usage{
convertToMatrix(bulkRegions)
}
\arguments{
\item{bulkRegions}{a dataframe of genomic coordinates of regions. Must contain either four columns
with names 'seqnames', 'start', 'end' and 'cluster' or two columns: 'sequence_name' and 'cluster'}
}
\value{
a data.table with sequence name as the first column and clusters as the following columns repressenting whether the sequence is in the cluster
}
\description{
\code{convertToMatrix} Reshape bulk annotation dataframe to a region by cluster matrix for permutation
}
