% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{downloadSampleData}
\alias{downloadSampleData}
\title{Download Sample Data}
\usage{
downloadSampleData(dest = "inst/extdata")
}
\arguments{
\item{dest}{folder to download data to}
}
\value{
destination folder
}
\description{
\code{downloadSampleData} Download sample data for use with vignette
}
