% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{findMarkers}
\alias{findMarkers}
\title{Find markers in scRNA-seq data}
\usage{
findMarkers(
  RNAMat,
  genes,
  barcodes,
  dims = 1:20,
  resolution = 0.5,
  only.pos = FALSE,
  min.cells = 3,
  scale.factor = 10000,
  nfeatures = 3000
)
}
\arguments{
\item{RNAMat}{gene-by-barcode raw count matrix of scRNASeq data}

\item{genes}{character vector of gene names}

\item{barcodes}{character vector of cell barcodes}

\item{dims}{integer vector of PCA dimensions to use}

\item{resolution}{numeric resolution to use in Louvain clustering}

\item{only.pos}{only return positive markers}

\item{min.cells}{used by Seurat, include features detected in at least this many cells}

\item{scale.factor}{used by Seurat, scale factor for cell-level normalization}

\item{nfeatures}{used by Seurat, Number of features to select as top variable features}
}
\value{
table of marker genes, labels, and logFC in expression
}
\description{
\code{findMarkers} Uses Seurat to find markers in scRNA-seq data and returns
a table of markers and labels that can be used by CellWalker. This implementation
uses default parameters and is designed for a simple first pass for creating
labels.
}
