% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctionsExtra.R
\name{findMotifs}
\alias{findMotifs}
\title{Identify motifs in given genomic regions}
\usage{
findMotifs(pRE)
}
\arguments{
\item{pRE}{a dataframe of genomic coordinates. Must contain at least three columns
with names 'seqnames', 'start' and 'end'.}
}
\value{
a dataframe with the first column pREs and second column motifs
}
\description{
\code{findMotifs} Identify motifs in given genomic regions
}
