% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{findUncertainLabels}
\alias{findUncertainLabels}
\title{Find uncertain labels}
\usage{
findUncertainLabels(
  cellWalk,
  cellTypes,
  threshold = 0.1,
  labelThreshold,
  plot = FALSE,
  normalize = FALSE
)
}
\arguments{
\item{cellWalk}{a cellWalk object}

\item{cellTypes}{character, vector of labels to use, all labels used by default}

\item{threshold}{numeric, quantile threshold for uncertain label}

\item{labelThreshold}{numeric, set a threshold below which cells aren't labeled (e.g. 0)}

\item{plot}{boolean, plot output matrix}

\item{normalize}{boolean, normalize plot scale}
}
\value{
cellWalk object with label uncertainty matrix (l-by-l) stored in "uncertaintyMatrix"
}
\description{
\code{findUncertainLabels()} generates an l-by-l matrix for how often each label is
confused for each other label at a given threshold, optinally normalized for total
counts of cells for each label (default FALSE)
}
\examples{
data("SampleCellWalkRData")
cellWalk <- findUncertainLabels(SampleCellWalkRData$cellWalk, threshold = .5)

}
