% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{getRegions}
\alias{getRegions}
\title{Get genomic regions}
\usage{
getRegions(geneBody = TRUE, genome = "hg38", names = "Entrez")
}
\arguments{
\item{geneBody}{boolean, use gene body or just promoter}

\item{genome}{which genome build (hg38, mm10)}

\item{names}{which gene names to use (Entrez, Ensembl)}
}
\value{
GRanges list of regions with gene_id
}
\description{
\code{getRegions} returns a GRanges object of how gene name identifiers map to
genomic regions either based on gene promoters or the full gene body
}
