% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labelGenesB}
\alias{labelGenesB}
\title{Label genes B}
\format{
A data frame with 1838 rows and 3 columns
}
\usage{
labelGenesB
}
\description{
A data.frame with labeling data
}
\keyword{datasets}
