% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{loadSampleData}
\alias{loadSampleData}
\title{Load Sample Data}
\usage{
loadSampleData()
}
\value{
sample data
}
\description{
\code{loadSampleData} Download sample data for use with vignette, only used for testing.
To load sampel data, simply call \code{data("SampleCellWalkRData")}
}
\examples{
loadSampleData()

}
