% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{mapArchRToGenes}
\alias{mapArchRToGenes}
\title{Map peaks to genes}
\usage{
mapArchRToGenes(labelGenes, ArchRproj, whichMat = "TileMatrix", regions)
}
\arguments{
\item{labelGenes}{data.frame with genes of interest in first column and corresponding
labels in second column}

\item{ArchRproj}{an ArchR project}

\item{whichMat}{string determing whether to use "TileMatrix" or "GeneScoreMatrix" from ArchR project}

\item{regions}{GRanges object of genomic regions associated with genes (only needed for "TileMatrix")}
}
\value{
list of genes and corresponding peaks by label
}
\description{
\code{mapArchRToGenes} Generates a mapping from genes to peaks per label
}
