% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCellTypesFunctions.R
\name{mapCellTypes}
\alias{mapCellTypes}
\title{Mapping cell type labels
\code{mapCellTypes} Random walk on the cells and cell type labels graph and permuted graphs, compute Z-scores for each pair of labels.
Cell type labels can be organized in a hierarchical structure.}
\usage{
mapCellTypes(
  cellGraph,
  labelEdgesList,
  labelEdgeWeights = NULL,
  wtrees = NULL,
  treeList = NULL,
  compute.Zscore = TRUE,
  nround = 50,
  groupsList = NULL,
  sampleDepth = 2000,
  ...
)
}
\arguments{
\item{cellGraph}{cell-cell graph, a cell-to-cell similarity matrix, row and colnames are cell barcodes}

\item{labelEdgeWeights}{a vector of the edge weight ratios between cell-label edges and cell-cell edges. One for each label set.
If it is NULL, will tune edge weights and get the optimal value for each  label set;
otherwise must have the same length as  \code{labelEdgeList}.}

\item{wtrees}{a dataframe/matrix the edge weight between nodes on the tree for each set of cell type labels. Each row is a label set. It must have two columns,
the first and second column is the edge weights traversing up/down the tree. If NULL, will set all edge weights to 1;
otherwise must have the same row as  \code{labelEdgeList}.}

\item{treeList}{a list of cell type trees, each is a phylo object of the hierarchical structure of a set of cell type labels. Cell type labels in labelEdgeList much contain all the tips on the tree.
If NULL, will not include hierarchical relationship between labels in the graph; otherwise must have the same length as \code{labelEdgeList}.}

\item{compute.Zscore}{whether to do permutation and compute Z-score, If not, will return influence score between labels.}

\item{nround}{rounds of permutations to compute null distribution. default: 50}

\item{groupsList}{a list of vectors showing groups of cells for each label set. Only permute edges between cells and labels within the same group of the cells. default: NULL, permute among all cells with all labels.
If not NULL, must have the same length as \code{labelEdgeList}. Each vector must have the same length as the rows of each labelEdge matrix. No permutation if group == 0.}

\item{sampleDepth}{subsample cells to for faster calculation to tune edge weight ratios (\code{labelEdgeWeights}), default: 2000}

\item{...}{other arguments pass to \code{tuneEdgeWeights} and \code{walkCells}}

\item{labelEdgeList}{a list of cell-to-label edges. Each is a matrix or dataframe of edges from a cell to each cell type label. Each row is a cell,
each column is a cell type label. The cell names of each matrix can be partially overlapped with the cells in \code{cellGraph}.}
}
\value{
A cellWalk2 object. if \code{compute.Zscore} is False a list including influence score between each set of cell types and labelEdgeWeights.
otherwise  a list including influence and Zscore between each set of cell types, as well as labelEdgeWeights.
}
\description{
Mapping cell type labels
\code{mapCellTypes} Random walk on the cells and cell type labels graph and permuted graphs, compute Z-scores for each pair of labels.
Cell type labels can be organized in a hierarchical structure.
}
