% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{mapPeaksToGenes}
\alias{mapPeaksToGenes}
\title{Map peaks to genes}
\usage{
mapPeaksToGenes(labelGenes, ATACMat, peaks, regions)
}
\arguments{
\item{labelGenes}{data.frame with genes of interest in first column and corresponding
labels in second column}

\item{ATACMat}{cell-by-peak matrix}

\item{peaks}{GRanges object of peaks (equal to number of rows in ATACMat)}

\item{regions}{GRanges object of genomic regions associated with genes}
}
\value{
list of genes and corresponding peaks by label
}
\description{
\code{mapPeaksToGenes} Generates a mapping from genes to peaks per label
}
\examples{
data("SampleCellWalkRData")
regions <- getRegions()
mapPeaksToGenes(SampleCellWalkRData$labelGenes,
                SampleCellWalkRData$ATACMat,
                SampleCellWalkRData$peaks,
                regions)

}
