% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{mapSnapATACToGenes}
\alias{mapSnapATACToGenes}
\title{Map peaks to genes}
\usage{
mapSnapATACToGenes(labelGenes, snap, whichMat = "bmat", regions)
}
\arguments{
\item{labelGenes}{data.frame with genes of interest in first column and corresponding
labels in second column}

\item{snap}{a snap object}

\item{whichMat}{string determing whether to use "bmat" or "gmat" from snap object}

\item{regions}{GRanges object of genomic regions associated with genes (only needed for "bmat")}
}
\value{
list of genes and corresponding peaks by label
}
\description{
\code{mapSnapATACToGenes} Generates a mapping from genes to peaks per label
}
