% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctionsExtra.R
\name{mergeATACSeq}
\alias{mergeATACSeq}
\title{Merge two ATACSeq data}
\usage{
mergeATACSeq(ATAC_Mat, peaks, ATAC_Mat2, peaks2)
}
\arguments{
\item{ATAC_Mat}{a peak-by-cell count matrix. Must have barcodes as colnames}

\item{peaks}{a dataframe of genomic coordinates. Must contain at least three columns
with names 'seqnames', 'start' and 'end'.}

\item{ATAC_Mat2}{another peak-by-cell count matrix. Must have cell barcodes as colnames different from colnames of \code{ATAC_Mat}}

\item{peaks2}{a dataframe of genomic coordinates of peaks the second count matrix. Must contain at least three columns
with names 'seqnames', 'start' and 'end'.}
}
\value{
a peak-by-cell matrix and a GRanges object of peaks
}
\description{
\code{mergeATACSeq} merge peaks from two ATACSeq data and combine cells to generate a single peak-by-cell matrix
}
