% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{mergeRNASeq}
\alias{mergeRNASeq}
\title{merge scRNAseq data}
\usage{
mergeRNASeq(
  RNAMatList,
  integrate = TRUE,
  min.cells = 3,
  scale.factor = 10000,
  nfeatures = 3000,
  ndim = 30,
  computeKNN = TRUE,
  knn = 20,
  ...
)
}
\arguments{
\item{RNAMatList}{a list of gene-by-barcode raw count matrix of scRNASeq data, either a matrix or a data.frame.
Each of them has cell barcodes as colnames and genes as rownames.}

\item{integrate}{Whether to use Seurat to remove batch effects and integrate all datasets provided. Default: True}

\item{min.cells}{used by Seurat, include features detected in at least this many cells.}

\item{scale.factor}{used by Seurat, scale factor for cell-level normalization}

\item{nfeatures}{used by Seurat, Number of features to select as top variable features and
integration feautres (if \code{integrate} is TRUE)}

\item{ndim}{integer vector of CCA dimensions to use for integration or PCA dimensions}

\item{computeKNN}{Whether to compute cell-cell graph using KNN. Default: True. If false, compute cell-cell similarity matrix based on PCs.}

\item{knn}{Defines k for the k-nearest neighbor algorithm}

\item{...}{other parameters can be passed to \code{CreateSeuratObject}}
}
\value{
normalized gene expression matrix and a matrix of cell-to-cell similarity matrix.
}
\description{
\code{mergeRNASeq} Uses Seurat to normalize raw count matrix of each data set, integrate and
generate cell-to-cell similarity graph that can be used by CellWalker.
}
