% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellWalkerFunctions.R
\name{normalizeInfluence}
\alias{normalizeInfluence}
\title{Normalize influence matrix}
\usage{
normalizeInfluence(labelCellInf)
}
\arguments{
\item{labelCellInf}{label-to-cell portion of influence matrix}
}
\value{
normalized influence matrix
}
\description{
\code{normalizeInfluence()} normalizes the label-to-cell portions of the
influence matrix for cross-label comparisons
}
