% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellWalkerFunctions.R
\name{PCAdist}
\alias{PCAdist}
\title{Compute distance in PCA space for matrix}
\usage{
PCAdist(m)
}
\arguments{
\item{m}{matrix}
}
\value{
Euclidian PCA distance matrix between rows of m
}
\description{
\code{PCAdist()} computes Euclidian PCA distance matrix between the rows
of a matrix, scaled by the greatest distance and subtracted from 1.
}
