% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{plotMultiLevelLabels}
\alias{plotMultiLevelLabels}
\title{Plot significant labels across heirarchy}
\usage{
plotMultiLevelLabels(cellWalk, labelScores, z = 2, whichBulk)
}
\arguments{
\item{cellWalk}{a cellWalk object}

\item{labelScores}{a matrix of label scores, output of labelBulk with allScores set to TRUE}

\item{z}{numeric z-score threshold for significance}

\item{whichBulk, }{numeric optionally identify which bulk peak to plot}
}
\value{
plot object p of either counts of signifcant labels or enrichment/depletion for single peak
}
\description{
\code{plotMultiLevelLabels()} Plots labels across heirarchy for bulk data by mapping them via
the calculated information matrix
}
