% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{plotZscoreDotplot}
\alias{plotZscoreDotplot}
\title{Plot Z-score}
\usage{
plotZscoreDotplot(Zscore, orderRow = T, orderCol = F, th = 3)
}
\arguments{
\item{Zscore}{a matrix or dataframe, rows are regions/cell type to be mapped to, and columns are cell types}

\item{orderRow}{whether to cluster bulk anntations based on Zscore profiles}

\item{cutoff}{the minimum Zscore to show on the plot and do clustering}
}
\value{
ggplot object
}
\description{
\code{plotZscoreDotplot} Plots Z-scores as dot plot
}
