% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{plotZscoreTree}
\alias{plotZscoreTree}
\title{Plot Z-score on a phylogeny}
\usage{
plotZscoreTree(celltree, Zscore, cutoff = 10, xrange = 20)
}
\arguments{
\item{celltree}{a phylo object, must have tip and node labels matched with column names of Zscores}

\item{Zscore}{a matrix or dataframe, rows are annotations/cell type to be mapped to, and columns are cell types on the tree, including both tips and internal nodes.}

\item{cutoff}{the minimum Zscore to show on the tree}

\item{xrange}{specify xlim in ggplot, adjust if the tree is partially missing from the figure}
}
\value{
ggplot object
}
\description{
\code{plotZscoreTree} Plots Z-scores as the color of branches of a phylogeny
}
