% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{processRNASeq}
\alias{processRNASeq}
\title{preprocess scRNAseq data}
\usage{
processRNASeq(
  RNAMat,
  meta.data = NULL,
  group.col = NULL,
  do.findMarkers = TRUE,
  computeKNN = TRUE,
  computeSimilarity = FALSE,
  knn = 20,
  buildTree = TRUE,
  genes = NULL,
  barcodes = NULL,
  dims = 1:20,
  resolution = 0.5,
  only.pos = FALSE,
  min.cells = 3,
  scale.factor = 10000,
  nfeatures = 3000,
  ...
)
}
\arguments{
\item{RNAMat}{gene-by-barcode raw count matrix of scRNASeq data, either a matrix or a data.frame}

\item{meta.data}{Additional cell-level metadata. Should be a data.frame where the rows are cell names matched with the column names of the counts matrix or \code{barcodes}
and the columns are additional metadata fields.}

\item{group.col}{the column name indicates the groups/cell clusters. If provided, will find markers for each group; otherwise, will use Seurat to identify clusters first.}

\item{do.findMarkers}{Whether to find cell type markers. Default: True.}

\item{computeKNN}{Whether to compute cell-cell graph using KNN. Default: True.}

\item{computeSimilarity}{Whether to compute cell-cell similarity matrix based on PCs. Default: False.}

\item{knn}{Defines k for the k-nearest neighbor algorithm}

\item{buildTree}{Whether to construct a cell type tree. Default: True}

\item{genes}{character vector of gene names. If NULL, use rownames of \code{RNAMat} as genes.}

\item{barcodes}{character vector of cell barcodes}

\item{dims}{integer vector of PCA dimensions to use}

\item{resolution}{numeric resolution to use in Louvain clustering}

\item{only.pos}{only return positive markers}

\item{min.cells}{used by Seurat, include features detected in at least this many cells.}

\item{scale.factor}{used by Seurat, scale factor for cell-level normalization}

\item{nfeatures}{used by Seurat, Number of features to select as top variable features}

\item{...}{other parameters can be passed to \code{CreateSeuratObject}}
}
\value{
normalized gene expression matrix,  table of marker genes, labels, and logFC in expression,
a matrix of cell-to-cell similarity matrix, a cell type tree and meta.data
}
\description{
\code{processRNASeq} Uses Seurat to normalize raw count matrix, generate cell-to-cell similarity graph and find markers in scRNA-seq data and returns
a table of markers and labels, cell-cell graph that can be used by CellWalker.
}
