% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellWalkerFunctions.R
\name{randomWalk}
\alias{randomWalk}
\title{Random walk with restarts}
\usage{
randomWalk(adj, r = 0.5, tensorflow = FALSE, steps)
}
\arguments{
\item{adj}{adjacency matrix}

\item{r}{restart probability}

\item{tensorflow}{boolean to indicate whether to compute on GPU}

\item{steps}{integer indicating number of steps to take if walk should not be run to convergence}
}
\value{
influence matrix, each column is the vector of influences on each row
}
\description{
\code{randomWalk()} solves a random walk with restarts on an adjacency matrix A using
the closed form solution for the influence matrix
\ifelse{html}{\out{F = r(I - (1 - r)W)<sup>-1</sup>}}{\eqn{F = r(I - (1 - r)W)^{-1}}}
where \ifelse{html}{\out{W = D<sup>-1</sup>A}}{\eqn{W = D^{-1}A}} and D is a diagonal
matrix of the sums of edge weights for each node and r is the restart probability
}
\examples{
data("SampleCellWalkRData")
adj <- combineGraph(SampleCellWalkRData$labelEdges, SampleCellWalkRData$cellEdges)
randomWalk(adj)

}
