% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleCellTypeTree}
\alias{SampleCellTypeTree}
\title{Cell type tree to test CellWalker2}
\format{
a phylo object
}
\usage{
SampleCellTypeTree
}
\description{
Cell type tree to test CellWalker2
}
\keyword{datasets}
