% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SampleCellWalkRData}
\alias{SampleCellWalkRData}
\title{List of 10 items used to test CellWalkR}
\format{
List of 10 items
}
\usage{
SampleCellWalkRData
}
\description{
10 items used for testing:
ATACMat
peaks
labelGenes
cellEdges
ATACGenePeak
labelEdges
cellWalk
filter
sampleEnhancers
labelScores
}
\keyword{datasets}
