% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleEnhancers}
\alias{sampleEnhancers}
\title{Enhancer regions}
\format{
A GRanges object with 20634 ranges
}
\usage{
sampleEnhancers
}
\description{
A GRanges object of enhancer regions
}
\keyword{datasets}
