% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{scoreDoublets}
\alias{scoreDoublets}
\title{Score doublets}
\usage{
scoreDoublets(cellWalk, plot = FALSE, embedding = "tSNE")
}
\arguments{
\item{cellWalk}{a cellWalk object}

\item{plot}{boolean, optionally plot doublet scores in embedding space}

\item{embedding}{method with which to embed data, either "tSNE" or "UMAP"}
}
\value{
cellWalk object with vector of doublet scores stored in "doubletScore"
}
\description{
\code{scoreDoublets()} generates an l-by-l matrix for how often each label is
confused for each other label at a given threshold, optinally normalized for total
counts of cells for each label (default FALSE)
}
\examples{
data("SampleCellWalkRData")
cellWalk <- scoreDoublets(SampleCellWalkRData$cellWalk)

}
