% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{selectLabels}
\alias{selectLabels}
\title{Select significant labels}
\usage{
selectLabels(labelScores, z = 2)
}
\arguments{
\item{labelScores}{a matrix of label scores, output of labelBulk with allScores set to TRUE}

\item{z}{numeric z-score threshold for significance}
}
\value{
list of signficant labels for each region in bulk data
}
\description{
\code{selectLabels()} Determines labels for bulk data by mapping them via
the calculated information matrix
}
