% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{selectMultiLevelLabels}
\alias{selectMultiLevelLabels}
\title{Select significant labels across heirarchy}
\usage{
selectMultiLevelLabels(cellWalk, labelScores, z = 2)
}
\arguments{
\item{cellWalk}{a cellWalk object}

\item{labelScores}{a matrix of label scores, output of labelBulk with allScores set to TRUE}

\item{z}{numeric z-score threshold for significance}
}
\value{
list of signficant labels for each region in bulk data
}
\description{
\code{selectMultiLevelLabels()} Determines labels across heirarchy for bulk data by mapping them via
the calculated information matrix
}
\examples{
data("SampleCellWalkRData")
cellWalk <- clusterLabels(SampleCellWalkRData$cellWalk)
\dontrun{selectMultiLevelLabels(cellWalk,SampleCellWalkRData$labelScores)}

}
