% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellWalkerFunctions.R
\name{sparseJaccard}
\alias{sparseJaccard}
\title{Compute Jaccard on sprase matrix}
\usage{
sparseJaccard(m)
}
\arguments{
\item{m}{sparse matrix}
}
\value{
Jaccard similarity matrix between columns of m
}
\description{
\code{sparseJaccard()} computes the Jaccard similarity between the columns
of a sparse matrix.
}
