% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellWalkerFunctions.R
\name{sparseJaccard0}
\alias{sparseJaccard0}
\title{Compute Jaccard on sprase matrix}
\usage{
sparseJaccard0(m)
}
\arguments{
\item{m}{sparse matrix}
}
\value{
Jaccard similarity matrix between rows of m
}
\description{
\code{sparseJaccard0()} computes the Jaccard similarity between the rows
of a sparse matrix.
}
\examples{
data("SampleCellWalkRData")
sparseJaccard0(SampleCellWalkRData$ATACMat)

}
