% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{storeBulk}
\alias{storeBulk}
\title{Store bulk label data}
\usage{
storeBulk(cellWalk, bulkPeaks, labelScores, bulkID)
}
\arguments{
\item{cellWalk}{a cellWalk object}

\item{bulkPeaks}{GRanges of peaks in bulk data or GRangesList of sets of peaks}

\item{labelScores}{a matrix of label scores, output of labelBulk with allScores set to TRUE}

\item{bulkID}{character string assigning name to bulk data}
}
\value{
cellWalk object with labels for each region in bulk data
}
\description{
\code{storeBulk()} stores labels for bulk data in the cellWalk object
}
\examples{
data("SampleCellWalkRData")
cellWalk <- storeBulk(SampleCellWalkRData$cellWalk,
                      SampleCellWalkRData$sampleEnhancers,
                      SampleCellWalkRData$labelScores)

}
