% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{storeMat}
\alias{storeMat}
\title{Store ATACMat data}
\usage{
storeMat(cellWalk, ATACMat, peaks)
}
\arguments{
\item{cellWalk}{a cellWalk object}

\item{ATACMat}{cell-by-peak matrix}

\item{peaks}{GRanges of peaks in ATACMat}
}
\value{
cellWalk object with original ATAC matrix stored
}
\description{
\code{storeMat()} stores the original ATAC matrix
}
\examples{
data("SampleCellWalkRData")
cellWalk <- storeMat(SampleCellWalkRData$cellWalk,
                     SampleCellWalkRData$ATACMat,
                     SampleCellWalkRData$peaks)

}
