% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellWalkerFunctions.R
\name{tensorJaccard}
\alias{tensorJaccard}
\title{Compute Jaccard on matrix using tensorflow}
\usage{
tensorJaccard(m)
}
\arguments{
\item{m}{sparse matrix}
}
\value{
Jaccard similarity matrix between rows of m
}
\description{
\code{tensorJaccard()} computes the Jaccard similarity between the rows
of a sparse matrix using tensorflow on a GPU
}
