% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{tuneEdgeWeights}
\alias{tuneEdgeWeights}
\title{Tune edge weights}
\usage{
tuneEdgeWeights(
  cellEdges,
  labelEdgesList,
  labelEdgeOpts = 10^seq(-4, 4, 1),
  sampleDepth = 10000,
  steps = Inf,
  cellTypes,
  parallel = FALSE,
  numCores = 1,
  trackProgress = FALSE,
  tensorflow = FALSE,
  method = "computeCellHomogeneity"
)
}
\arguments{
\item{cellEdges}{matrix of cell-to-cell similarity (c-by-c)}

\item{labelEdgesList}{list of cell-to-label similarity matrices each with dimensions c-by-l}

\item{labelEdgeOpts}{vector of weights to test for label edges}

\item{sampleDepth}{integer depth to subsample cells to for faster calculation}

\item{steps}{integer indicating number of steps to take if walk should not be run to convergence}

\item{cellTypes}{list of names of cell types, if not provided unique cell labels are used}

\item{parallel}{boolean, execute in parallel}

\item{numCores}{integer, number of cores to use for parallel execution}

\item{trackProgress}{boolean, print percent run completed}

\item{tensorflow}{boolean to indicate whether to compute on GPU}

\item{method}{criterion to choose the optimal parameters. If annotate cells or map cell types, use computeCellHomogeneity otherwise use  computeLabelHomogeneity}
}
\value{
data frame of weights with corresponding cell homogeneity
}
\description{
\code{tuneEdgeWeights} generates cell homogeneity scores for possible edge
weight settings across all labelEdge matrices. To speed up calculations cells
can be subsampled to a specified sampleDepth (default 10000).
}
\examples{
data("SampleCellWalkRData")
labelEdgesList <- list(SampleCellWalkRData$labelEdges)
tuneEdgeWeights(SampleCellWalkRData$cellEdges, labelEdgesList, 10^seq(1,7,1))

}
