% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{walkCells}
\alias{walkCells}
\title{Run CellWalker Method}
\usage{
walkCells(
  cellEdges,
  labelEdgesList,
  labelEdgeWeights,
  sampleDepth,
  tensorflow = FALSE
)
}
\arguments{
\item{cellEdges}{matrix of cell-to-cell similarity (c-by-c)}

\item{labelEdgesList}{list of cell-to-label similarity matrices each with dimensions c-by-l}

\item{labelEdgeWeights}{vector of weights to test for label edges}

\item{sampleDepth}{depth to subsample cells to for faster calculation}

\item{tensorflow}{boolean to indicate whether to compute on GPU}
}
\value{
cellWalk object with influence matrix and labels
}
\description{
\code{walkCells} applies the CellWalker method to the given cell-to-cell and
cell-to-label edges.
}
