% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataParsingFunctions.R
\name{computeCellSim}
\alias{computeCellSim}
\title{Compute Cell Similarity}
\usage{
computeCellSim(ATACMat, method = "Jaccard")
}
\arguments{
\item{ATACMat}{either a cell-by-peak matrix, a SnapATAC object, an ArchR project, or a Cicero count table}

\item{method}{method used to compute cell similarity}
}
\value{
matrix of cell-to-cell similarity
}
\description{
\code{computeCellSim} computes a cell-to-cell similarity matrix using the given method.
Currently only Jaccard similarity is implemented.
}
