% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractiveFunctions.R
\name{launchViz}
\alias{launchViz}
\title{Launch Interactve Visualization}
\usage{
launchViz(cellWalk)
}
\arguments{
\item{cellWalk}{a cellWalk object}
}
\description{
\code{launchViz()} launches an interactve visualization
}
