% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownstreamAnalysisFunctions.R
\name{plotCells}
\alias{plotCells}
\title{Plot Cells}
\usage{
plotCells(
  cellWalk,
  cellTypes,
  labelThreshold,
  initial_dims = 10,
  perplexity = 50,
  recompute = FALSE,
  plot = TRUE,
  seed
)
}
\arguments{
\item{cellWalk}{a cellWalk object}

\item{cellTypes}{character, optional vector of labels to use, all labels used by default. Most likely label among those listed
will be used. If only a single label is provided, all cells will be colored by their score for that label. If two labels are
given, the difference in score for each cell is shown.}

\item{labelThreshold}{numeric, set a threshold below which cells aren't labeled (e.g. 0)}

\item{initial_dims}{numeric, number of PCA dims to use for tSNE}

\item{perplexity}{numeric, perplexity parameter for tSNE}

\item{recompute}{boolean, recompute tSNE}

\item{plot}{boolean, optionally don't plot output and only compute the embedding}

\item{seed}{numeric, random seed}
}
\value{
cellWalk object with embedding stored in "tSNE"
}
\description{
\code{plotCells()} generates a tSNE plot based on cell-to-cell influence
}
